+++ src/builder.rs
@@ -1,6 +1,7 @@
 //! Logic for building Orchard components of transactions.
 
 use std::convert::TryFrom;
+use std::fmt;
 use std::iter;
 
 use ff::Field;
@@ -363,9 +364,9 @@ impl Builder {
 }
 
 /// Marker trait representing bundle signatures in the process of being created.
-pub trait InProgressSignatures {
+pub trait InProgressSignatures: fmt::Debug {
     /// The authorization type of an Orchard action in the process of being authorized.
-    type SpendAuth;
+    type SpendAuth: fmt::Debug;
 }
 
 /// Marker for a bundle in the process of being built.
@@ -375,7 +376,7 @@ pub struct InProgress<P, S: InProgressSignatures> {
     sigs: S,
 }
 
-impl<P, S: InProgressSignatures> Authorization for InProgress<P, S> {
+impl<P: fmt::Debug, S: InProgressSignatures> Authorization for InProgress<P, S> {
     type SpendAuth = S::SpendAuth;
 }
 
@@ -488,7 +489,7 @@ impl MaybeSigned {
     }
 }
 
-impl<P, V> Bundle<InProgress<P, Unauthorized>, V> {
+impl<P: fmt::Debug, V> Bundle<InProgress<P, Unauthorized>, V> {
     /// Loads the sighash into this bundle, preparing it for signing.
     ///
     /// This API ensures that all signatures are created over the same sighash.
@@ -534,7 +535,7 @@ impl<V> Bundle<InProgress<Proof, Unauthorized>, V> {
     }
 }
 
-impl<P, V> Bundle<InProgress<P, PartiallyAuthorized>, V> {
+impl<P: fmt::Debug, V> Bundle<InProgress<P, PartiallyAuthorized>, V> {
     /// Signs this bundle with the given [`SpendAuthorizingKey`].
     ///
     /// This will apply signatures for all notes controlled by this spending key.
@@ -679,9 +680,9 @@ pub mod testing {
             for note in notes.iter() {
                 let leaf = MerkleHashOrchard::from_cmx(&note.commitment().into());
                 tree.append(&leaf);
-                tree.witness();
+                let (position, leaf) = tree.witness().expect("tree is not empty");
 
-                let path = tree.authentication_path(&leaf).unwrap().into();
+                let path = MerklePath::from((position, tree.authentication_path(position, &leaf).expect("we just witnessed the path")));
                 notes_and_auth_paths.push((*note, path));
             }