+++ src/tree.rs
@@ -19,7 +19,7 @@ use serde::de::{Deserializer, Error};
 use serde::ser::Serializer;
 use serde::{Deserialize, Serialize};
 use std::iter;
-use subtle::{Choice, ConditionallySelectable, ConstantTimeEq, CtOption};
+use subtle::{Choice, ConditionallySelectable, CtOption};
 
 // The uncommitted leaf is defined as pallas::Base(2).
 // <https://zips.z.cash/protocol/protocol.pdf#thmuncommittedorchard>
@@ -160,7 +160,7 @@ impl MerklePath {
 
 /// A newtype wrapper for leaves and internal nodes in the Orchard
 /// incremental note commitment tree.
-#[derive(Copy, Clone, Debug)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord)]
 pub struct MerkleHashOrchard(pallas::Base);
 
 impl MerkleHashOrchard {
@@ -187,25 +187,6 @@ impl MerkleHashOrchard {
     /// a non-canonical encoding.
     pub fn from_bytes(bytes: &[u8; 32]) -> CtOption<Self> {
         pallas::Base::from_repr(*bytes).map(MerkleHashOrchard)
-    }
-}
-
-/// This instance should only be used for hash table key comparisons.
-impl std::cmp::PartialEq for MerkleHashOrchard {
-    fn eq(&self, other: &Self) -> bool {
-        self.0.ct_eq(&other.0).into()
-    }
-}
-
-/// This instance should only be used for hash table key comparisons.
-impl std::cmp::Eq for MerkleHashOrchard {}
-
-/// This instance should only be used for hash table key hashing.
-impl std::hash::Hash for MerkleHashOrchard {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        <Option<pallas::Base>>::from(self.0)
-            .map(|b| b.to_repr())
-            .hash(state)
     }
 }
 
@@ -302,7 +283,7 @@ pub mod testing {
         {
             let cmx = MerkleHashOrchard::from_bytes(&tv.leaves[i]).unwrap();
             tree.append(&cmx);
-            tree.witness();
+            tree.witness().expect("tree is not empty");
 
             assert_eq!(tree.root().0, pallas::Base::from_repr(tv.root).unwrap());
 
@@ -312,14 +293,13 @@ pub mod testing {
             for j in 0..=i {
                 let leaf = MerkleHashOrchard::from_bytes(&tv.leaves[j]).unwrap();
                 assert_eq!(
-                    tree.authentication_path(&leaf),
-                    Some((
-                        j.try_into().unwrap(),
+                    tree.authentication_path(j.try_into().unwrap(), &leaf),
+                    Some(
                         tv.paths[j]
                             .iter()
                             .map(|v| MerkleHashOrchard::from_bytes(v).unwrap())
                             .collect()
-                    ))
+                    )
                 );
             }
         }