+++ CHANGELOG.md
@@ -9,13 +9,18 @@ and this project adheres to Rust's notion of
 ### Added
 - `orchard::keys`:
   - `DiversifierIndex::to_bytes`
+  - `FullViewingKey::derive_internal`
   - `IncomingViewingKey::diversifier_index`
+- `orchard::note`:
+  - `impl PartialEq, Eq, PartialOrd, Ord for Nullifier`
 - `orchard::primitives::redpallas::VerificationKey::verify`
-- `orchard::tree::MerklePath::from_parts`
+- `orchard::tree`:
+  - `MerklePath::from_parts`
+  - `impl PartialEq, Eq, PartialOrd, Ord for MerkleHashOrchard`
 - `impl From<orchard::bundle::BundleCommitment> for [u8; 32]`
 
 ### Changed
-- MSRV is now 1.54.0.
+- MSRV is now 1.56.1.
 - Bumped dependencies to `pasta_curves 0.3`.
 - The following methods now have an additional `rng: impl RngCore` argument:
   - `orchard::builder::Bundle::create_proof`
@@ -25,6 +30,8 @@ and this project adheres to Rust's notion of
   instead of `i64: From<&'a V>`.
 - `orchard::Bundle::binding_validating_key` now requires the bound
   `V: Into<i64>` instead of `V: Into<ValueSum>`.
+- `orchard::builder::InProgressSignatures` and `orchard::bundle::Authorization`
+  now have `Debug` bounds on themselves and their associated types.
 
 ### Removed
 - `orchard::bundle`:
@@ -34,8 +41,10 @@ and this project adheres to Rust's notion of
 - `orchard::keys`:
   - `FullViewingKey::default_address`
   - `IncomingViewingKey::default_address`
-- `DiversifierKey` (use the APIs on `FullViewingKey` and `IncomingViewingKey`
-  instead).
+  - `DiversifierKey` (use the APIs on `FullViewingKey` and `IncomingViewingKey`
+    instead).
+- `impl std::hash::Hash for orchard::tree::MerkleHashOrchard` (use `BTreeMap`
+  instead of `HashMap`).
 - `orchard::value::ValueSum::from_raw`
 
 ## [0.1.0-beta.1] - 2021-12-17